requirements.txt
python-telegram-bot==20.6
Pillow==10.4.0
imagehash==4.3.1
pytesseract==0.3.10
openpyxl==3.1.5
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
